inline global constant references, move statements to their own file
